DjangoBooks/requirements.txt
alembic==1.0.0
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
diff-match-patch==20121119
Django==1.11
django-celery==3.2.2
django-celery-results==1.0.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-formtools==2.1
django-import-export==1.0.1
django-redis==4.9.0
django-reversion==2.0.13
djangorestframework==3.8.2
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
httpie==0.9.9
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==3.0.37
lxml==4.2.3
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.0.0
protobuf==3.6.0
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.8.1
pypi==2.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
WMI==1.4.9
WTForms==2.2.1
xlrd==1.1.0
xlwt==1.3.0
